backend/resume-service/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
spacy
python-docx
